pyproject.toml
[tool.poetry]
name = "homework-drf"
version = "0.1.0"
description = ""
authors = ["Hilshat76"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.3"
psycopg2-binary = "^2.9.10"
pillow = "^11.0.0"
python-dotenv = "^1.0.1"
black = "^24.10.0"
djangorestframework = "^3.15.2"
django-filter = "^24.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
